% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Say hello to someone}
\usage{
say_hello(someone = "Berra")
}
\arguments{
\item{someone}{Character. The name of the person you want to say hello to}
}
\value{
Side effect: A message in the console
}
\description{
Say hello to someone
}
\examples{
say_hello() # Hello Berra
say_hello(someone = "Seb") # Hello Seb

hello_anna <- say_hello(someone = "Anna") 
hello_anna
# library(ggplot2)
# ggplot() + geom_point()
}
